% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTA.R
\name{addTA}
\alias{addTA}
\title{Add Technical Analysis (TA) chart with OHLC + Volume}
\usage{
addTA(
  ohlc_data,
  volume_data = NULL,
  name = "Symbol",
  id = "ohlc",
  title = "OHLC Chart",
  subtitle = NULL,
  height = 600
)
}
\arguments{
\item{ohlc_data}{list of OHLC values: list(list(ms, open, high, low, close), ...)}

\item{volume_data}{list of volume values: list(list(ms, volume), ...). Varsayılan NULL.}

\item{name}{character, OHLC serisi adı (varsayılan "Symbol")}

\item{id}{character, OHLC serisinin id'si (varsayılan "ohlc")}

\item{title}{chart başlığı}

\item{subtitle}{chart alt başlığı}

\item{height}{chart yüksekliği (ör. 600 veya "600px")}
}
\value{
highchartS htmlwidget
}
\description{
Bu fonksiyon, OHLC (candlestick) ve opsiyonel Volume serilerini birlikte çizer.
Kullanımı kolaylaştırmak için \code{highchartS()} fonksiyonunu çağırır.
}
\examples{
# Örnek OHLC ve Volume verisi
ohlc_data <- list(
  list(1735689600000, 93576, 94509.41, 93489.03, 94401.14),
  list(1735693200000, 94401.13, 94408.72, 93578.77, 93607.74)
)

volume_data <- list(
  list(1735689600000, 755.99),
  list(1735693200000, 586.53)
)

# Grafik çizimi
addTA(
  ohlc_data   = ohlc_data,
  volume_data = volume_data,
  name = "AAPL",
  id   = "aapl",
  title = "AAPL Historical",
  subtitle = "With OHLC + Volume"
)

}
