% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartS.R
\name{as_hc_ohlc}
\alias{as_hc_ohlc}
\title{OHLC veri hazırlama yardımcı fonksiyonu (opsiyonel)}
\usage{
as_hc_ohlc(df, name = "OHLC")
}
\arguments{
\item{df}{data.frame ya da xts, kolonlar: time, open, high, low, close}

\item{name}{Seri adı}
}
\description{
OHLC veri hazırlama yardımcı fonksiyonu (opsiyonel)
}
