% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartS.R
\name{as_hc_series}
\alias{as_hc_series}
\title{xts/zoo veya data.frame'i Highcharts dizi formatına çevirin}
\usage{
as_hc_series(x, name = "Series")
}
\arguments{
\item{x}{xts/zoo (tek kolon) ya da data.frame(Date|POSIXct, value)}

\item{name}{Seri adı}
}
\description{
Highcharts zaman ekseni milisaniye bekler. Bu fonksiyon, tek kolonlu xts/zoo
veya (Date|POSIXct, value) iki kolonlu data.frame'i
list(list(ms, val), ...) yapısına çevirir.
}
