% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartS.R
\name{highchartS}
\alias{highchartS}
\title{highchartS: Highstock tabanlı htmlwidget (highcharter'sız)}
\usage{
highchartS(
  series,
  title = "My Highstock Chart",
  yAxis = NULL,
  rangeSelector = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{series}{list of series; her biri list(name=?, data=?), data: list( list(ms, val), ... )}

\item{title}{(character) grafik başlığı}

\item{yAxis}{(list) Highcharts/Highstock yAxis konfig (tek ya da çoklu)}

\item{rangeSelector}{(logical) Highstock range selector açık/kapalı}

\item{width, height, elementId}{htmlwidget boyut/ID}
}
\description{
highchartS: Highstock tabanlı htmlwidget (highcharter'sız)
}
